+++ chess/calc.hpp
@@ -14,17 +14,17 @@ class seen_positions {
 public:
 	seen_positions();
 
-	unsigned long long pos[100 + MAX_DEPTH + MAX_QDEPTH + 10]; // Must be at least 50 full moves + max depth and add some safety margin.
+	uint64_t pos[100 + MAX_DEPTH + MAX_QDEPTH + 10]; // Must be at least 50 full moves + max depth and add some safety margin.
 	int root_position; // Index of root position in seen_positions
 	int null_move_position;
 
-	bool is_two_fold( unsigned long long hash, int ply ) const;
-	bool is_three_fold( unsigned long long hash, int ply ) const;
+	bool is_two_fold( uint64_t hash, int ply ) const;
+	bool is_three_fold( uint64_t hash, int ply ) const;
 };
 
 struct new_best_move_callback
 {
-	virtual void on_new_best_move( position const& p, color::type c, int depth, int evaluation, unsigned long long nodes, pv_entry const* pv );
+	virtual void on_new_best_move( position const& p, color::type c, int depth, int evaluation, uint64_t nodes, pv_entry const* pv );
 };
 extern new_best_move_callback default_new_best_move_callback;
 
@@ -36,7 +36,7 @@ public:
 
 	// May modify seen_positions at indexes > root_position
 	bool calc( position& p, color::type c, move& m, int& res,
-		   unsigned long long move_time_limit, unsigned long long time_remaining, int clock,
+		   uint64_t move_time_limit, uint64_t time_remaining, int clock,
 		   seen_positions& seen, short last_mate,
 		   new_best_move_callback& new_best_cb = default_new_best_move_callback );
 
@@ -86,6 +86,6 @@ public:
 };
 
 // Depth is number of plies to search multiplied by depth_factor
-short step( int depth, int ply, context& ctx, position const& p, unsigned long long hash, int current_evaluation, color::type c, check_map const& check, short alpha, short beta, pv_entry* pv, bool last_was_null );
+short step( int depth, int ply, context& ctx, position const& p, uint64_t hash, int current_evaluation, color::type c, check_map const& check, short alpha, short beta, pv_entry* pv, bool last_was_null );
 
 #endif
