+++ chess/eval.hpp
@@ -68,7 +68,7 @@ short evaluate_full( position const& p, color::type c, short eval_fast );
 
 short evaluate_move( position const& p, color::type c, short current_evaluation, move const& m, position::pawn_structure& outPawns );
 
-void evaluate_pawns( unsigned long long const white_pawns, unsigned long long const black_pawns, short* eval );
+void evaluate_pawns( uint64_t const white_pawns, uint64_t const black_pawns, short* eval );
 
 short get_material_value( pieces::type pi );
